Move base handler back into exposition.
