+++ tests/test_exposition.py
@@ -13,7 +13,7 @@ from prometheus_client import Gauge, Counter, Summary, Histogram, Metric
 from prometheus_client import CollectorRegistry, generate_latest
 from prometheus_client import push_to_gateway, pushadd_to_gateway, delete_from_gateway
 from prometheus_client import CONTENT_TYPE_LATEST, instance_ip_grouping_key
-from prometheus_client.handlers.base import handler as default_handler
+from prometheus_client.exposition import default_handler
 from prometheus_client.handlers.basic_auth import handler as basic_auth_handler
 
 try:
